Changed decimal in INetwork, INode, IWeightedNode as well as implementations; implemented Node with TDD

+++ NeuralBotMasterFramework/NeuralBotMasterFramework/Interfaces/INetwork.cs
@@ -14,6 +14,6 @@ namespace NeuralBotMasterFramework.Interfaces
 
         void SetInput();
         void Propagate();
-        decimal[] GetOutput();
+        double[] GetOutput();
     }
 }
